+++ mobile/android/android-components/components/browser/engine-gecko/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngineSession.kt
@@ -884,6 +884,10 @@ class GeckoEngineSession(
                 notifyObservers { onMetaViewportFitChanged(layoutInDisplayCutoutMode) }
             }
         }
+
+        override fun onShowDynamicToolbar(geckoSession: GeckoSession) {
+            notifyObservers { onShowDynamicToolbar() }
+        }
     }
 
     private fun createContentBlockingDelegate() = object : ContentBlocking.Delegate {