+++ mobile/android/android-components/components/concept/engine/src/test/java/mozilla/components/concept/engine/EngineSessionTest.kt
@@ -80,6 +80,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onLoadRequest("https://www.mozilla.org", true, true) }
         session.notifyInternalObservers { onLaunchIntentRequest("https://www.mozilla.org", null) }
         session.notifyInternalObservers { onProcessKilled() }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
 
         verify(observer).onLocationChange("https://www.mozilla.org")
         verify(observer).onLocationChange("https://www.firefox.com")
@@ -113,6 +114,7 @@ class EngineSessionTest {
         verify(observer).onLoadRequest("https://www.mozilla.org", true, true)
         verify(observer).onLaunchIntentRequest("https://www.mozilla.org", null)
         verify(observer).onProcessKilled()
+        verify(observer).onShowDynamicToolbar()
         verifyNoMoreInteractions(observer)
     }
 
@@ -150,6 +152,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onCrash() }
         session.notifyInternalObservers { onLoadRequest("https://www.mozilla.org", true, true) }
         session.notifyInternalObservers { onLaunchIntentRequest("https://www.mozilla.org", null) }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
         session.unregister(observer)
 
         val mediaSessionController: MediaSession.Controller = mock()
@@ -186,6 +189,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onCrash() }
         session.notifyInternalObservers { onLoadRequest("https://www.mozilla.org", true, true) }
         session.notifyInternalObservers { onLaunchIntentRequest("https://www.firefox.com", null) }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
 
         verify(observer).onLocationChange("https://www.mozilla.org")
         verify(observer).onProgress(25)
@@ -207,6 +211,7 @@ class EngineSessionTest {
         verify(observer).onCrash()
         verify(observer).onLoadRequest("https://www.mozilla.org", true, true)
         verify(observer).onLaunchIntentRequest("https://www.mozilla.org", null)
+        verify(observer).onShowDynamicToolbar()
         verify(observer, never()).onLocationChange("https://www.firefox.com")
         verify(observer, never()).onProgress(100)
         verify(observer, never()).onLoadingStateChange(false)
@@ -270,6 +275,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onCancelContentPermissionRequest(permissionRequest) }
         session.notifyInternalObservers { onAppPermissionRequest(permissionRequest) }
         session.notifyInternalObservers { onWindowRequest(windowRequest) }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
 
         session.unregisterObservers()
 
@@ -304,6 +310,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onMediaPositionStateChanged(mediaSessionPositionState) }
         session.notifyInternalObservers { onMediaMuteChanged(true) }
         session.notifyInternalObservers { onMediaFullscreenChanged(true, mediaSessionElementMetadata) }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
 
         verify(observer).onLocationChange("https://www.mozilla.org")
         verify(observer).onProgress(25)
@@ -322,6 +329,7 @@ class EngineSessionTest {
         verify(observer).onContentPermissionRequest(permissionRequest)
         verify(observer).onCancelContentPermissionRequest(permissionRequest)
         verify(observer).onWindowRequest(windowRequest)
+        verify(observer).onShowDynamicToolbar()
         verify(observer, never()).onLocationChange("https://www.firefox.com")
         verify(observer, never()).onProgress(100)
         verify(observer, never()).onLoadingStateChange(false)
@@ -405,6 +413,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onCancelContentPermissionRequest(permissionRequest) }
         session.notifyInternalObservers { onAppPermissionRequest(permissionRequest) }
         session.notifyInternalObservers { onWindowRequest(windowRequest) }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
 
         session.close()
 
@@ -439,6 +448,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onMediaPositionStateChanged(mediaSessionPositionState) }
         session.notifyInternalObservers { onMediaMuteChanged(true) }
         session.notifyInternalObservers { onMediaFullscreenChanged(true, mediaSessionElementMetadata) }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
 
         verify(observer).onLocationChange("https://www.mozilla.org")
         verify(observer).onProgress(25)
@@ -457,6 +467,7 @@ class EngineSessionTest {
         verify(observer).onContentPermissionRequest(permissionRequest)
         verify(observer).onCancelContentPermissionRequest(permissionRequest)
         verify(observer).onWindowRequest(windowRequest)
+        verify(observer).onShowDynamicToolbar()
         verify(observer, never()).onLocationChange("https://www.firefox.com")
         verify(observer, never()).onProgress(100)
         verify(observer, never()).onLoadingStateChange(false)
@@ -527,6 +538,7 @@ class EngineSessionTest {
         otherSession.notifyInternalObservers { onMediaPositionStateChanged(mediaSessionPositionState) }
         otherSession.notifyInternalObservers { onMediaMuteChanged(true) }
         otherSession.notifyInternalObservers { onMediaFullscreenChanged(true, mediaSessionElementMetadata) }
+        otherSession.notifyInternalObservers { onShowDynamicToolbar() }
         verify(observer, never()).onLocationChange("https://www.mozilla.org")
         verify(observer, never()).onProgress(25)
         verify(observer, never()).onLoadingStateChange(true)
@@ -552,6 +564,7 @@ class EngineSessionTest {
         verify(observer, never()).onMediaPositionStateChanged(mediaSessionPositionState)
         verify(observer, never()).onMediaMuteChanged(true)
         verify(observer, never()).onMediaFullscreenChanged(true, mediaSessionElementMetadata)
+        verify(observer, never()).onShowDynamicToolbar()
 
         session.notifyInternalObservers { onLocationChange("https://www.mozilla.org") }
         session.notifyInternalObservers { onProgress(25) }
@@ -578,6 +591,7 @@ class EngineSessionTest {
         session.notifyInternalObservers { onMediaPositionStateChanged(mediaSessionPositionState) }
         session.notifyInternalObservers { onMediaMuteChanged(true) }
         session.notifyInternalObservers { onMediaFullscreenChanged(true, mediaSessionElementMetadata) }
+        session.notifyInternalObservers { onShowDynamicToolbar() }
         verify(observer, times(1)).onLocationChange("https://www.mozilla.org")
         verify(observer, times(1)).onProgress(25)
         verify(observer, times(1)).onLoadingStateChange(true)
@@ -603,6 +617,7 @@ class EngineSessionTest {
         verify(observer, times(1)).onMediaPositionStateChanged(mediaSessionPositionState)
         verify(observer, times(1)).onMediaMuteChanged(true)
         verify(observer, times(1)).onMediaFullscreenChanged(true, mediaSessionElementMetadata)
+        verify(observer, times(1)).onShowDynamicToolbar()
         verifyNoMoreInteractions(observer)
     }
 
@@ -833,6 +848,7 @@ class EngineSessionTest {
         defaultObserver.onCancelContentPermissionRequest(mock(PermissionRequest::class.java))
         defaultObserver.onWindowRequest(mock(WindowRequest::class.java))
         defaultObserver.onCrash()
+        defaultObserver.onShowDynamicToolbar()
     }
 
     @Test
@@ -935,6 +951,7 @@ class EngineSessionTest {
         observer.onLoadRequest("https://www.mozilla.org", true, true)
         observer.onLaunchIntentRequest("https://www.mozilla.org", null)
         observer.onProcessKilled()
+        observer.onShowDynamicToolbar()
     }
 }