+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarAutocompleteFeatureTest.kt
@@ -121,6 +121,14 @@ class ToolbarAutocompleteFeatureTest {
         override fun disableScrolling() {
             fail()
         }
+
+        override fun collapse() {
+            fail()
+        }
+
+        override fun expand() {
+            fail()
+        }
     }
 
     @Test