+++ mobile/android/android-components/components/feature/toolbar/src/test/java/mozilla/components/feature/toolbar/ToolbarInteractorTest.kt
@@ -104,6 +104,14 @@ class ToolbarInteractorTest {
         override fun disableScrolling() {
             fail()
         }
+
+        override fun collapse() {
+            fail()
+        }
+
+        override fun expand() {
+            fail()
+        }
     }
 
     @Test